feat: replace routes with /demo/* & /board
