--- src/About.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const About = () => <div>about</div>
-
-export default About
