--- src/Home.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-function Home() {
-	return <div>home</div>
-}
-
-export default Home
