--- src/Users.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const Users = () => <div>users</div>
-
-export default Users
